+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -1,5 +1,16 @@
 ## Release History
 
+### 4.68.0-beta.1 (Unreleased)
+
+#### Features Added
+
+#### Breaking Changes
+
+#### Bugs Fixed
+* Fixed applications possibly not closing gracefully due to thread `partition-availability-staleness-check` not being a daemon thread. - [PR 44441](https://github.com/Azure/azure-sdk-for-java/pull/44441).
+
+#### Other Changes
+
 ### 4.67.0 (2025-02-20)
 
 #### Bugs Fixed