+++ sdk/eventgrid/azure-messaging-eventgrid-systemevents/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 1.0.0-beta.1 (2025-02-24)
+## 1.0.0-beta.2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0-beta.1 (2025-02-26)
 
 ### Features Added
 - Azure SystemEvents client library for Java. This package contains Microsoft Azure SystemEvents client library.